requirements.txt
# Core test automation stack
playwright==1.48.0         # Browser automation for Chromium, Firefox, WebKit
pytest==8.3.3              # Test runner
pytest-playwright==0.5.0   # Pytest plugin for Playwright integration

# Process-safe locking for state validation
filelock==3.15.4

# Optional but recommended utilities
pytest-html==4.1.1         # Generates HTML test reports
pytest-xdist==3.6.1        # Parallel test execution
faker==30.3.0              # Fake data generation for test inputs
python-dotenv==1.0.1       # Load environment variables from .env files